Ben Commit Changes

+++ src/server/scripts/Commands/cs_modify.cpp
@@ -1240,7 +1240,9 @@ data.WriteByteSeq(guid[4]);
         if (!*args)
             return false;
 
-        uint16 display_id = (uint16)atoi((char*)args);
+        //BENDEV: Fix uint16 to uint32 to hold bigger display ID's from later expacs
+        //uint16 display_id = (uint16)atoi((char*)args);
+        uint32 display_id = (uint32)atoi((char*)args);
 
         Unit* target = handler->getSelectedUnit();
         if (!target)
